+++ src/Molecules/FlexTable/Cell/TextWrapper.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
-import styled from 'styled-components';
 import { Typography } from '../../..';
 import { TextWrapperComponent } from './Cell.types';
 import { getFontSizeTypographyType } from '../shared/textUtils';
-
-const StyledTypography = styled(Typography)`
-  display: inline-block;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-`;
+import { StyledTruncateTooltip } from '../shared';
 
 export const TextWrapper: TextWrapperComponent = ({ fontSize = 'm', children }) => (
-  <StyledTypography type={getFontSizeTypographyType(fontSize)} color={t => t.color.text}>
-    {children}
-  </StyledTypography>
+  <StyledTruncateTooltip label={children}>
+    <Typography type={getFontSizeTypographyType(fontSize)} color={t => t.color.text}>
+      {children}
+    </Typography>
+  </StyledTruncateTooltip>
 );