+++ src/Molecules/FlexTable/FlexTable.tsx
@@ -4,8 +4,8 @@ import { HeaderRow, FooterRow, Row } from './Row';
 import { Header } from './Header';
 import { Footer } from './Footer';
 import { Cell } from './Cell';
-import { constants, ColumnProvider } from './shared';
-import { Props, FlexTableComponents } from './FlexTable.types';
+import { constants, ColumnProvider, CellInlineContainer } from './shared';
+import { FlexTableComponents, FlexTableComponent } from './FlexTable.types';
 import { FlexTableProvider, useFlexTable } from './shared/FlexTableProvider';
 import { ExpandCell } from './Cell/ExpandCell';
 import { Typography } from '../..';
@@ -41,7 +41,7 @@ const StyledTypography = styled(Typography)`
   padding-left: ${p => p.theme.spacing.unit(1)}px;
 `;
 
-const FlexTable: React.FC<Props> & FlexTableComponents = ({
+const FlexTable: FlexTableComponent & FlexTableComponents = ({
   className,
   density = 'm',
   expandable = false,
@@ -49,6 +49,10 @@ const FlexTable: React.FC<Props> & FlexTableComponents = ({
   children,
   title,
   fontSize = 'm',
+  sm,
+  md,
+  lg,
+  xl,
   ...htmlProps
 }) => (
   <FlexTableProvider
@@ -56,6 +60,10 @@ const FlexTable: React.FC<Props> & FlexTableComponents = ({
     stickyHeader={stickyHeader}
     fontSize={fontSize}
     expandable={expandable}
+    sm={sm}
+    md={md}
+    lg={lg}
+    xl={xl}
   >
     {/* pass sticky with context instead of prop-drilling, since context might change */}
     <FlexTableContainer className={className} {...htmlProps}>
@@ -74,6 +82,7 @@ FlexTable.Header = Header;
 FlexTable.Footer = Footer;
 FlexTable.ExpandCell = ExpandCell;
 FlexTable.Cell = Cell;
+FlexTable.CellInlineContainer = CellInlineContainer;
 FlexTable.CONSTANTS = constants;
 
 export default FlexTable;