+++ src/Molecules/FlexTable/Header/HeaderContent/TextWrapper.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Typography } from '../../../..';
 import { getFontSizeTypographyType } from '../../shared/textUtils';
 import { TextWrapperComponent } from './HeaderContent.types';
-
-const StyledTypography = styled(Typography).withConfig({
-  shouldForwardProp: p => !['density'].includes(p),
-})`
-  display: inline-block;
-  word-break: break-word;
-`;
+import { Typography } from '../../../..';
+import { StyledTruncateTooltip } from '../../shared';
 
 export const TextWrapper: TextWrapperComponent = ({ fontSize = 'm', sorted, children }) => (
-  <StyledTypography
-    type={getFontSizeTypographyType(fontSize)}
-    color={t => (sorted ? t.color.text : t.color.label)}
-    weight={sorted ? 'bold' : 'regular'}
-  >
-    {children}
-  </StyledTypography>
+  <StyledTruncateTooltip label={children}>
+    <Typography
+      type={getFontSizeTypographyType(fontSize)}
+      color={t => (sorted ? t.color.text : t.color.label)}
+      weight={sorted ? 'bold' : 'regular'}
+    >
+      {children}
+    </Typography>
+  </StyledTruncateTooltip>
 );