+++ src/Molecules/FlexTable/Row/Row.tsx
@@ -1,15 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import { ExpandProps, RowComponent, RowComponents } from './Row.types';
-import { Box, Flexbox, Button, Icon } from '../../../index';
+import { RowComponent, RowComponents } from './Row.types';
+import { Flexbox } from '../../..';
 import { ColorFn } from '../../../common/Types/sharedTypes';
 import { getDensityPaddings } from '../shared/textUtils';
 import { Density } from '../shared/shared.types';
 import { useFlexTable } from '../shared/FlexTableProvider';
-import { ExpandItems, ExpandItem } from './ExpandItems';
-import { ExpandCell } from '../Cell/ExpandCell';
-import { COLUMN_ID_EXPAND, ICON_COLUMN_DEFAULT_FLEX_PROPS } from '../shared/constants';
-import FlexTable from '../FlexTable';
+import { ExpandItems, ExpandItem } from './components/ExpandItems';
+import { RenderForSizes } from '../shared';
+import { ExpandElement, ExpandArea } from './components';
 
 /* the cells are padded by row gutter 1 unit (4px) */
 const StyledRow = styled(Flexbox).withConfig({
@@ -32,10 +31,13 @@ const StyledRow = styled(Flexbox).withConfig({
 }>`
   ${p =>
     !p.hideSeparator && !p.expanded ? `border-bottom: 1px solid ${p.separatorColor(p.theme)}` : ''};
+
   padding-right: ${p => (p.expandable ? p.theme.spacing.unit(2) : p.theme.spacing.unit(1))}px;
   padding-left: ${p => (p.expandable ? p.theme.spacing.unit(1.5) : p.theme.spacing.unit(0.5))}px;
+
   border-left: ${p => p.theme.spacing.unit(0.5)}px solid
-    ${p => (p.expanded ? p.theme.color.cta : 'transparent')};
+    ${p => (p.expanded && p.expandable ? p.theme.color.cta : 'transparent')};
+
   ${p =>
     p.hoverHighlight &&
     !p.expanded &&
@@ -45,6 +47,19 @@ const StyledRow = styled(Flexbox).withConfig({
 
   padding-top: ${p => getDensityPaddings(p.density)}px;
   padding-bottom: ${p => getDensityPaddings(p.density)}px;
+
+  margin-right: 0;
+  margin-left: 0;
+
+  & > * {
+    &:first-child {
+      padding-left: 0;
+    }
+
+    &:last-child {
+      padding-right: 0;
+    }
+  }
 `;
 
 const StyledExpandedRow = styled('div').withConfig({
@@ -54,62 +69,32 @@ const StyledExpandedRow = styled('div').withConfig({
   border-bottom: 1px solid ${p => p.separatorColor(p.theme)};
 `;
 
-export const ExpandButton: React.FC<{ expanded: boolean; onClick: () => void }> = ({
-  expanded,
-  onClick,
-}) => (
-  <Button variant="neutral" onClick={onClick} aria-expanded={expanded}>
-    {expanded ? <Icon.ChevronUp /> : <Icon.ChevronDown />}
-  </Button>
-);
-
-const ExpandElement: React.FC<
-  ExpandProps & { expandable: boolean; isContent: boolean; setExpand: (expanded: boolean) => void }
-> = ({
-  expandable,
-  isContent,
-  expanded = false,
-  onExpandToggle,
-  setExpand,
-  expandChildren,
-  expandItems,
-}) => {
-  if (!expandable) {
-    return null;
-  }
-  if (!isContent) {
-    return (
-      <FlexTable.Header
-        columnId={FlexTable.CONSTANTS.COLUMN_ID_EXPAND}
-        {...ICON_COLUMN_DEFAULT_FLEX_PROPS}
-      />
-    );
-  }
-
-  return (
-    <ExpandCell
-      columnId={COLUMN_ID_EXPAND}
-      expanded={expanded}
-      onClick={() => (onExpandToggle ? onExpandToggle(!expanded) : setExpand(!expanded))}
-      disabled={!(expandChildren || expandItems)}
-    />
-  );
-};
-
 const Row: RowComponent & RowComponents = ({
   className,
   expanded = false,
   hoverHighlight = true,
   hideSeparator = false,
+  // If false means that it's a header
   isContent = true,
   separatorColor = theme => theme.color.divider,
   onExpandToggle,
-  expandChildren,
-  expandItems,
+  expandChildren: expandChildrenXs,
+  expandItems: expandItemsXs,
   children,
+  sm,
+  md,
+  lg,
+  xl,
   ...htmlProps
 }) => {
-  const { density, expandable } = useFlexTable();
+  const {
+    density: xsDensity,
+    expandable: xsExpandable,
+    sm: smTable,
+    md: mdTable,
+    lg: lgTable,
+    xl: xlTable,
+  } = useFlexTable();
   const [expand, setExpand] = useState(expanded);
 
   useEffect(() => {
@@ -118,43 +103,68 @@ const Row: RowComponent & RowComponents = ({
 
   return (
     <>
-      <StyledRow
-        container
-        alignItems="center"
-        hoverHighlight={hoverHighlight}
-        className={className}
-        hideSeparator={hideSeparator}
-        role="row"
-        expanded={expand}
-        separatorColor={separatorColor}
-        density={density}
-        expandable={expandable}
-        {...htmlProps}
-      >
-        {children}
-        <ExpandElement
-          isContent={isContent}
-          expanded={expand}
-          expandItems={expandItems}
-          expandChildren={expandChildren}
-          onExpandToggle={onExpandToggle}
-          setExpand={setExpand}
-          expandable={expandable}
-        />
-      </StyledRow>
+      <RenderForSizes
+        xs={{ density: xsDensity, expandable: xsExpandable }}
+        sm={smTable}
+        md={mdTable}
+        lg={lgTable}
+        xl={xlTable}
+        Container={({ density, expandable, children: component }) => (
+          <StyledRow
+            container
+            alignItems="center"
+            hoverHighlight={hoverHighlight}
+            className={className}
+            hideSeparator={hideSeparator}
+            role="row"
+            expanded={expand}
+            separatorColor={separatorColor}
+            density={density}
+            expandable={expandable}
+            gutter={1}
+            {...htmlProps}
+          >
+            {component}
+          </StyledRow>
+        )}
+        Component={({ expandable }) => (
+          <>
+            {children}
+            {expandable && (
+              <ExpandElement
+                isContent={isContent}
+                expanded={expand}
+                onExpandToggle={onExpandToggle}
+                disabled={!expandChildrenXs && !expandItemsXs}
+                setExpand={setExpand}
+              />
+            )}
+          </>
+        )}
+      />
 
       {expand && (
-        <StyledExpandedRow role="row" separatorColor={separatorColor}>
-          {/* TODO should we rather have padding specified in ExpandItems? */}
-          <Box px={5} pb={2} md={{ px: 5, pt: 5, pb: 0 }} role="cell">
-            {expandItems && <ExpandItems items={expandItems} />}
-            {expandChildren && (
-              <Box pt={2} md={{ px: 5, pt: 0, pb: 5 }}>
-                {expandChildren}
-              </Box>
-            )}
-          </Box>
-        </StyledExpandedRow>
+        <RenderForSizes
+          xs={{
+            expandItems: expandItemsXs,
+            expandChildren: expandChildrenXs,
+          }}
+          sm={sm}
+          md={md}
+          lg={lg}
+          xl={xl}
+          Container={({ children: component }) => (
+            <StyledExpandedRow role="row" separatorColor={separatorColor}>
+              {component}
+            </StyledExpandedRow>
+          )}
+          Component={({ expandChildren, expandItems }) => {
+            if (expandItems.length === 0 && !expandChildren) {
+              return null;
+            }
+            return <ExpandArea expandItems={expandItems} expandChildren={expandChildren} />;
+          }}
+        />
       )}
     </>
   );