+++ src/Molecules/FlexTable/shared/FlexTableProvider/FlexTableProvider.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { FlexTableProviderComponent, FlexTableState } from './FlexTableProvider.types';
+import { FlexTableProviderComponent, FlexTableContextProps } from './FlexTableProvider.types';
 
-export const FlexTableContext = React.createContext<FlexTableState | undefined>(undefined);
+export const FlexTableContext = React.createContext<FlexTableContextProps>(undefined);
 
 export const FlexTableProvider: FlexTableProviderComponent = ({
   children,
@@ -9,8 +9,14 @@ export const FlexTableProvider: FlexTableProviderComponent = ({
   expandable,
   fontSize,
   stickyHeader,
+  sm,
+  md,
+  lg,
+  xl,
 }) => (
-  <FlexTableContext.Provider value={{ density, stickyHeader, fontSize, expandable }}>
+  <FlexTableContext.Provider
+    value={{ density, stickyHeader, fontSize, expandable, sm, md, lg, xl }}
+  >
     {children}
   </FlexTableContext.Provider>
 );